p2p: migrate mdns logic
