+++ p2p/host_test.go
@@ -1,0 +1,335 @@
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
+package p2p
+
+import (
+	"reflect"
+	"testing"
+	"time"
+)
+
+// tests host connect method
+func TestConnect(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeB, _, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	addrInfoB, err := nodeB.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoB)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	peerCountA := nodeA.host.peerCount()
+	peerCountB := nodeB.host.peerCount()
+
+	if peerCountA != 1 {
+		t.Error(
+			"node A does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountA,
+		)
+	}
+
+	if peerCountB != 1 {
+		t.Error(
+			"node B does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountB,
+		)
+	}
+}
+
+// tests host bootstrap method on start
+func TestBootstrap(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	addrA := nodeA.host.fullAddr()
+
+	configB := &Config{
+		BootstrapNodes: []string{addrA.String()},
+		Port:           7002,
+		RandSeed:       2,
+		NoGossip:       true,
+		NoMdns:         true,
+	}
+
+	nodeB, _, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	peerCountA := nodeA.host.peerCount()
+	peerCountB := nodeB.host.peerCount()
+
+	if peerCountA != 1 {
+		t.Error(
+			"node A does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountA,
+		)
+	}
+
+	if peerCountB != 1 {
+		t.Error(
+			"node B does not have expected peer count",
+			"\nexpected:", 1,
+			"\nreceived:", peerCountB,
+		)
+	}
+}
+
+// tests host ping method
+func TestPing(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeB, _, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	addrInfoB, err := nodeB.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoB)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.ping(addrInfoB.ID)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	addrInfoA, err := nodeA.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeB.host.ping(addrInfoA.ID)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+// tests host send method
+func TestSend(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeB, msgSendB, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	addrInfoB, err := nodeB.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoB)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.send(addrInfoB.ID, testMessage)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	select {
+	case msg := <-msgSendB:
+		if !reflect.DeepEqual(msg, testMessage) {
+			t.Error(
+				"node B received unexpected message from node A",
+				"\nexpected:", testMessage,
+				"\nreceived:", msg,
+			)
+		}
+	case <-time.After(TestMessageTimeout):
+		t.Error("node B timeout waiting for message from node A")
+	}
+}
+
+// tests host broadcast method
+func TestBroadcast(t *testing.T) {
+	configA := &Config{
+		Port:        7001,
+		RandSeed:    1,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeA, _, _ := createTestService(t, configA)
+	defer nodeA.Stop()
+
+	nodeA.host.noStatus = true
+
+	addrA := nodeA.host.fullAddrs()[0]
+
+	configB := &Config{
+		Port:        7002,
+		RandSeed:    2,
+		NoBootstrap: true,
+		NoGossip:    true,
+		NoMdns:      true,
+	}
+
+	nodeB, msgSendB, _ := createTestService(t, configB)
+	defer nodeB.Stop()
+
+	nodeB.host.noStatus = true
+
+	addrInfoB, err := nodeB.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoB)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	configC := &Config{
+		BootstrapNodes: []string{addrA.String()},
+		Port:           7003,
+		RandSeed:       3,
+		NoGossip:       true,
+		NoMdns:         true,
+	}
+
+	nodeC, msgSendC, _ := createTestService(t, configC)
+	defer nodeC.Stop()
+
+	nodeC.host.noStatus = true
+
+	addrInfoC, err := nodeC.host.addrInfo()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = nodeA.host.connect(*addrInfoC)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	nodeA.host.broadcast(testMessage)
+
+	select {
+	case msg := <-msgSendB:
+		if !reflect.DeepEqual(msg, testMessage) {
+			t.Error(
+				"node B received unexpected message from node A",
+				"\nexpected:", testMessage,
+				"\nreceived:", msg,
+			)
+		}
+	case <-time.After(TestMessageTimeout):
+		t.Error("node B timeout waiting for message")
+	}
+
+	select {
+	case msg := <-msgSendC:
+		if !reflect.DeepEqual(msg, testMessage) {
+			t.Error(
+				"node C received unexpected message from node A",
+				"\nexpected:", testMessage,
+				"\nreceived:", msg,
+			)
+		}
+	case <-time.After(TestMessageTimeout):
+		t.Error("node C timeout waiting for message")
+	}
+
+}
